+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a7f63be8ea353e64c89d59a4a528f697
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
