+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Animation/Enemy.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a07e83c24bef77c43aa1ad415fb7af4c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
