+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Animation/EnemySurvey.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a1197932df4abf41ab7d62b4d10e870
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
