+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Scripts/DebugOnly.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4548711d1951cb44189f27211e78656f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
