+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Scripts/DebugOnly/EnemyRouteDrawer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4534ff3dce383d249a7970562d11d6ba
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
