+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Scripts/State.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b6cd558300de7624bb8321082535af82
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
