+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Scripts/State/EnemyChaser.cs
@@ -6,7 +6,7 @@ using UnityEngine.AI;
 /// <summary>
 /// プレイヤーの追跡を行うエネミー
 /// </summary>
-public class EnemyChaser : MonoBehaviour
+public class EnemyChaser : StateMachineBehaviour
 {
     // ナビメッシュ
     [SerializeField]
@@ -23,7 +23,7 @@ public class EnemyChaser : MonoBehaviour
     /// <summary>
     /// 開始
     /// </summary>
-    void OnEnable()
+    public override void OnStateEnter(Animator animator, AnimatorStateInfo animatorStateInfo, int layerIndex)
     {
         // 移動スピードをセット
         navMeshAgent.speed = moveSpeed;
@@ -32,7 +32,7 @@ public class EnemyChaser : MonoBehaviour
     /// <summary>
     /// 更新
     /// </summary>
-    void Update()
+    public override void OnStateUpdate(Animator animator, AnimatorStateInfo animatorStateInfo, int layerIndex)
     {
         // 追跡対象の位置をセット
         navMeshAgent.SetDestination(player.position);
