+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Scripts/State/EnemySearch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d8979b15d1bf26542b5c321adad5777a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
