+++ pro_builder_test/Assets/ProgrammerSample/Mitsumaru/Scripts/State/EnemySurvey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b4e2d1a0a1d92049ab54b4479ff587f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
